Add REST_FRAMEWORK variable

+++ generators/rest_framework/index.js
@@ -35,6 +35,12 @@ module.exports = class extends Generator {
     writing() {
         this._add_in_installed_apps('rest_framework');
         this._add_in_url_patterns(`path('${this.currentConfig.apiUrlPath}/', include('rest_framework.urls', namespace='rest_framework'))`);
+        this._add_variable_in_settings(`
+        REST_FRAMEWORK = {
+            'DEFAULT_PERMISSION_CLASSES': [
+                'rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly'
+            ]
+        }`);
     }
 
     _add_in_installed_apps(appName) {
@@ -46,6 +52,18 @@ module.exports = class extends Generator {
 
         const content = this.fs.read(settingsFilePath);
         const updated = content.replace(/INSTALLED_APPS = \[/, `INSTALLED_APPS = [\n    '${appName}',`);
+        this.fs.write(settingsFilePath, updated);
+    }
+
+    _add_variable_in_settings(data) {
+        const settingsFilePath = this.destinationPath(path.join(this.currentConfig.projectName, 'settings.py'));
+        if (!this.fs.exists(settingsFilePath)) {
+            this.log('Cannot file settings.py');
+            return;
+        }
+
+        const content = this.fs.read(settingsFilePath);
+        const updated = content.concat('\n', data, '\n');
         this.fs.write(settingsFilePath, updated);
     }
